Simplify wire codec boundaries
